Split up classes in Item.cs
